Removed some features
Removed “Forgot Password” and “Change Details” features.

+++ MyCred.php
@@ -31,12 +31,15 @@ function regUser() {
 	$formData['password']=hashPass($formData['password']);
 
 	// Connect to database
-	$dbAccess=new dbControl($dbConnect['host'],$dbConnect['database'],$dbConnect['user'],$dbConnect['password']);
+	if (($dbAccess=new dbControl($dbConnect['host'],$dbConnect['database'],$dbConnect['user'],$dbConnect['password']))==null) {
+		$formError['message']="Error connecting to database! Please try again.";
+		return;
+	}
 
 	// Check if user account exists and if not, register user
 	if (!$dbAccess->emailExists($formData['email'])) {
 		// Attempt to login user
-		if ($dbAccess->regUserDb()) {
+		if ($dbAccess->regUser()) {
 			// Establish Session and login to Account Actions Menu
 			$dbAccess->closeDb();
 			$_SESSION['email']=$formData['email'];
@@ -68,7 +71,10 @@ function loginUser() {
 	}
 
 	// Connect to database to check user exists and compare users password
-	$dbAccess=new dbControl($dbConnect['host'],$dbConnect['database'],$dbConnect['user'],$dbConnect['password']);
+	if (($dbAccess=new dbControl($dbConnect['host'],$dbConnect['database'],$dbConnect['user'],$dbConnect['password']))==null) {
+		$formError['message']="Error connecting to database! Please try again.";
+		return;
+	}
 
 	// Check if user account exists and check password
 	$sqlrecords=$dbAccess->checkPassword($formData['email']);
@@ -80,6 +86,8 @@ function loginUser() {
 
 	// Check user entry exists
 	if (count($sqlrecords)==0) {
+		$formError['message']="Incorrect Username or Password!";
+		$dbAccess->closeDb();
 		return; // No email match found.
 	}
 
@@ -107,7 +115,10 @@ function chgPassword() {
 	$oldPassword=clean_input($_POST['oldPassword']);
 
 	// Connect to database to check users old password is valid and to set the new password
-	$dbAccess=new dbControl($dbConnect['host'],$dbConnect['database'],$dbConnect['user'],$dbConnect['password']);
+	if (($dbAccess=new dbControl($dbConnect['host'],$dbConnect['database'],$dbConnect['user'],$dbConnect['password']))==null) {
+		$formError['message']="Error connecting to database! Please try again.";
+		return;
+	}
 
 	// Check if old password matches current password
 	$sqlrecords=$dbAccess->checkPassword($_SESSION['email']);
@@ -151,8 +162,12 @@ function delAccount() {
 	global $dbConnect;
 
 	// Connect to database so users account can be deleted from it
-	$dbAccess=new dbControl($dbConnect['host'],$dbConnect['database'],$dbConnect['user'],$dbConnect['password']);
+	if (($dbAccess=new dbControl($dbConnect['host'],$dbConnect['database'],$dbConnect['user'],$dbConnect['password']))==null) {
+		$formError['message']="Error connecting to database! Please try again.";
+		return;
+	}
 
+	// Delete user entry from database
 	if (!$dbAccess->deleteUser($_SESSION['email'])) {
 		$dbAccess->closeDb();
 		header("Location: http://" . $_SERVER['SERVER_NAME'] . ":" . $_SERVER['SERVER_PORT'] . "/MyCred/index.php");
